refactor(sspi): Kerberos client implementation (#435)
