+++ src/kerberos/client/generators.rs
@@ -68,9 +68,9 @@ const MD5_CHECKSUM_TYPE: [u8; 1] = [0x07];
 // https://www.rfc-editor.org/rfc/rfc4120#section-5.4.1
 pub const DEFAULT_AS_REQ_OPTIONS: [u8; 4] = [0x00, 0x81, 0x00, 0x10];
 
-// Renewable, Canonicalize, Enc-tkt-in-skey are on by default
+// Renewable, Canonicalize.
 // https://www.rfc-editor.org/rfc/rfc4120#section-5.4.1
-const DEFAULT_TGS_REQ_OPTIONS: [u8; 4] = [0x00, 0x81, 0x00, 0x08];
+const DEFAULT_TGS_REQ_OPTIONS: [u8; 4] = [0x00, 0x81, 0x00, 0x00];
 
 const DEFAULT_PA_PAC_OPTIONS: [u8; 4] = [0x40, 0x00, 0x00, 0x00];
 
@@ -83,8 +83,8 @@ pub const AUTHENTICATOR_DEFAULT_CHECKSUM: [u8; 24] = [
     0x00, 0x00, 0x00, 0x00, 0x00,
 ];
 
-// [MS-KILE] 3.3.5.6.1 Client Principal Lookup
-// https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-kile/6435d3fb-8cf6-4df5-a156-1277690ed59c
+/// [MS-KILE] 3.3.5.6.1 Client Principal Lookup
+/// https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-kile/6435d3fb-8cf6-4df5-a156-1277690ed59c
 pub fn get_client_principal_name_type(username: &str, _domain: &str) -> u8 {
     if username.contains('@') {
         NT_ENTERPRISE
@@ -143,11 +143,16 @@ fn matches_domain(domain: &str, mapping_domain: &str) -> bool {
     }
 }
 
+/// Parameters for generating pa-datas for [AsReq] message.
 #[derive(Debug)]
 pub struct GenerateAsPaDataOptions<'a> {
     pub password: &'a str,
+    /// Salt for deriving the encryption key.
+    ///
+    /// The salt value should be extracted from the [KrbError] message.
     pub salt: Vec<u8>,
     pub enc_params: EncryptionParams,
+    /// Flag that indicates whether to generate pa-datas.
     pub with_pre_auth: bool,
 }
 
@@ -210,6 +215,7 @@ pub fn generate_pa_datas_for_as_req(options: &GenerateAsPaDataOptions) -> Result
     Ok(pa_datas)
 }
 
+/// Parameters for generating [AsReq].
 #[derive(Debug)]
 pub struct GenerateAsReqOptions<'a> {
     pub realm: &'a str,
@@ -295,13 +301,18 @@ pub fn generate_as_req(pa_datas: Vec<PaData>, kdc_req_body: KdcReqBody) -> AsReq
     })
 }
 
+/// Parameters for generating [TgsReq].
 #[derive(Debug)]
 pub struct GenerateTgsReqOptions<'a> {
     pub realm: &'a str,
     pub service_principal: &'a str,
     pub session_key: &'a [u8],
+    /// [Ticket] extracted from the [AsRep] message.
     pub ticket: Ticket,
+    /// [Authenticator] to be included in [TgsReq] pa-data.
     pub authenticator: &'a mut Authenticator,
+    /// If the Kerberos U2U auth is negotiated, then this parameter must have one ticket: TGT ticket of the application service.
+    /// Otherwise, set it to `None`.
     pub additional_tickets: Option<Vec<Ticket>>,
     pub enc_params: &'a EncryptionParams,
     pub context_requirements: ClientRequestFlags,
@@ -329,6 +340,9 @@ pub fn generate_tgs_req(options: GenerateTgsReqOptions) -> Result<TgsReq> {
     let mut tgs_req_options = KdcOptions::from_bits(u32::from_be_bytes(DEFAULT_TGS_REQ_OPTIONS)).unwrap();
     if context_requirements.contains(ClientRequestFlags::DELEGATE) {
         tgs_req_options |= KdcOptions::FORWARDABLE;
+    }
+    if context_requirements.contains(ClientRequestFlags::USE_SESSION_KEY) {
+        tgs_req_options |= KdcOptions::ENC_TKT_IN_SKEY;
     }
 
     let req_body = KdcReqBody {
@@ -526,22 +540,33 @@ pub struct AuthenticatorChecksumExtension {
     pub extension_value: Vec<u8>,
 }
 
+/// Encryption key.
 #[derive(Debug)]
 pub struct EncKey {
+    /// Encryption type.
     pub key_type: CipherSuite,
+    /// Encryption key value.
     pub key_value: Vec<u8>,
 }
 
+/// Input parameters for generating ApReq Authenticator.
 #[derive(Debug)]
 pub struct GenerateAuthenticatorOptions<'a> {
+    /// [KdcRep] from previous interaction with KDC.
     pub kdc_rep: &'a KdcRep,
+    /// Sequence number.
     pub seq_num: Option<u32>,
+    /// Sub-session encryption key.
     pub sub_key: Option<EncKey>,
+    /// Authenticator checksum options.
     pub checksum: Option<ChecksumOptions>,
+    /// Channel bindings.
     pub channel_bindings: Option<&'a ChannelBindings>,
+    /// Possible authenticator extensions.
     pub extensions: Vec<AuthenticatorChecksumExtension>,
 }
 
+/// Generated ApReq Authenticator.
 #[instrument(level = "trace", ret)]
 pub fn generate_authenticator(options: GenerateAuthenticatorOptions) -> Result<Authenticator> {
     let GenerateAuthenticatorOptions {
@@ -723,13 +748,13 @@ pub fn generate_ap_req(
     }))
 }
 
-// returns supported authentication types
+/// Returns supported authentication types.
 pub fn get_mech_list() -> MechTypeList {
     MechTypeList::from(vec![MechType::from(oids::ms_krb5()), MechType::from(oids::krb5())])
 }
 
 pub fn generate_neg_token_init(username: &str, service_name: &str) -> Result<ApplicationTag0<GssApiNegInit>> {
-    let krb5_neg_token_init: ApplicationTag<_, 0> = ApplicationTag::from(KrbMessage {
+    let krb5_neg_token_init = ApplicationTag0(KrbMessage {
         krb5_oid: ObjectIdentifierAsn1::from(oids::krb5_user_to_user()),
         krb5_token_id: TGT_REQ_TOKEN_ID,
         krb_msg: TgtReq {
