+++ tests/sspi/client_server/kerberos/kdc.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 use std::hash::{Hash, Hasher};
 
+use picky_asn1::bit_string::BitString;
 use picky_asn1::date::GeneralizedTime;
 use picky_asn1::restricted_string::IA5String;
 use picky_asn1::wrapper::{
@@ -15,17 +16,17 @@ use picky_krb::constants::key_usages::{
 };
 use picky_krb::constants::types::{
     AS_REP_MSG_TYPE, KRB_ERROR_MSG_TYPE, PA_ENC_TIMESTAMP, PA_ENC_TIMESTAMP_KEY_USAGE, PA_ETYPE_INFO2_TYPE,
-    PA_TGS_REQ_TYPE, TGS_REP_MSG_TYPE,
+    PA_TGS_REQ_TYPE, TGS_REP_MSG_TYPE, TGT_REP_MSG_TYPE,
 };
 use picky_krb::crypto::CipherSuite;
 use picky_krb::data_types::{
-    Authenticator, EncTicketPart, EncryptedData, EncryptionKey, EtypeInfo2Entry, KerberosStringAsn1, KerberosTime,
-    LastReq, LastReqInner, Microseconds, PaData, PaEncTsEnc, PrincipalName, Realm, Ticket, TicketInner,
+    Authenticator, EncTicketPart, EncryptedData, EncryptionKey, EtypeInfo2Entry, KerberosFlags, KerberosStringAsn1,
+    KerberosTime, LastReq, LastReqInner, Microseconds, PaData, PaEncTsEnc, PrincipalName, Realm, Ticket, TicketInner,
     TransitedEncoding,
 };
 use picky_krb::messages::{
     ApReq, ApReqInner, AsRep, AsReq, EncAsRepPart, EncKdcRepPart, EncTgsRepPart, KdcRep, KdcReq, KdcReqBody, KrbError,
-    KrbErrorInner, TgsRep, TgsReq,
+    KrbErrorInner, TgsRep, TgsReq, TgtRep, TgtReq,
 };
 use rand::rngs::OsRng;
 use rand::{Rng, RngCore};
@@ -59,13 +60,32 @@ impl Hash for UserName {
     }
 }
 
+/// Collection of validators of incoming Kerberos messages.
+///
+/// The user can provide customs validators to check messages generated by our Kerberos implementation.
+pub struct Validators {
+    /// Closure that validates the incoming AsReq message.
+    pub as_req: Box<dyn Fn(&AsReq) + Send + Sync>,
+    /// Closure that validates the incoming TgsReq message.
+    pub tgs_req: Box<dyn Fn(&TgsReq) + Send + Sync>,
+}
+
+impl Default for Validators {
+    fn default() -> Self {
+        Self {
+            as_req: Box::new(|_| {}),
+            tgs_req: Box::new(|_| {}),
+        }
+    }
+}
+
 /// Simple mock of the KDC server.
 ///
 /// We use it to test our Kerberos implementation.
 /// This KDC implementation performs only small amount of all possible checks on
 /// the incoming Kerberos messages: encryption keys + key usage number usage
 /// and some mandatory fields like `pa-datas`.
-/// All other validations like checking user/service names should be done separately.
+/// All other validations like checking user/service names should be done separately. See [Validators] structure for more details.
 pub struct KdcMock {
     /// Domain's Kerberos realm.
     realm: String,
@@ -73,12 +93,24 @@ pub struct KdcMock {
     keys: HashMap<UserName, Vec<u8>>,
     /// Represents users credentials.
     users: HashMap<UserName, PasswordCreds>,
+    /// Incoming Kerberos messages validators.
+    validators: Validators,
 }
 
 impl KdcMock {
     /// Creates a new [KdcMock].
-    pub fn new(realm: String, keys: HashMap<UserName, Vec<u8>>, users: HashMap<UserName, PasswordCreds>) -> Self {
-        Self { realm, keys, users }
+    pub fn new(
+        realm: String,
+        keys: HashMap<UserName, Vec<u8>>,
+        users: HashMap<UserName, PasswordCreds>,
+        validators: Validators,
+    ) -> Self {
+        Self {
+            realm,
+            keys,
+            users,
+            validators,
+        }
     }
 
     fn make_err<const ERROR_CODE: u32>(sname: PrincipalName, realm: Realm, salt: Option<String>) -> KrbError {
@@ -185,10 +217,65 @@ impl KdcMock {
         Ok(key)
     }
 
+    fn make_ticket(
+        realm: KerberosStringAsn1,
+        session_key: Vec<u8>,
+        service_key: &[u8],
+        kdc_options: KerberosFlags,
+        sname: PrincipalName,
+        cname: PrincipalName,
+    ) -> Ticket {
+        let auth_time = OffsetDateTime::now_utc();
+        let end_time = auth_time + Duration::days(1);
+
+        let ticket_enc_part = EncTicketPart {
+            flags: ExplicitContextTag0::from(kdc_options.clone()),
+            key: ExplicitContextTag1::from(EncryptionKey {
+                key_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![18])),
+                key_value: ExplicitContextTag1::from(OctetStringAsn1::from(session_key)),
+            }),
+            crealm: ExplicitContextTag2::from(realm.clone()),
+            cname: ExplicitContextTag3::from(cname),
+            transited: ExplicitContextTag4::from(TransitedEncoding {
+                // the client is unable to check these fields, so we can put any values we want
+                tr_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![0])),
+                contents: ExplicitContextTag1::from(OctetStringAsn1::from(vec![1])),
+            }),
+            auth_time: ExplicitContextTag5::from(KerberosTime::from(GeneralizedTime::from(auth_time))),
+            starttime: Optional::from(None),
+            endtime: ExplicitContextTag7::from(KerberosTime::from(GeneralizedTime::from(end_time))),
+            renew_till: Optional::from(None),
+            caddr: Optional::from(None),
+            authorization_data: Optional::from(None),
+        };
+
+        let ticket_enc_data = CipherSuite::Aes256CtsHmacSha196
+            .cipher()
+            .encrypt(
+                service_key,
+                TICKET_REP,
+                &picky_asn1_der::to_vec(&ticket_enc_part).unwrap(),
+            )
+            .unwrap();
+
+        Ticket::from(TicketInner {
+            tkt_vno: ExplicitContextTag0::from(IntegerAsn1::from(vec![KERBEROS_VERSION])),
+            realm: ExplicitContextTag1::from(realm),
+            sname: ExplicitContextTag2::from(sname),
+            enc_part: ExplicitContextTag3::from(EncryptedData {
+                etype: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
+                kvno: Optional::from(None),
+                cipher: ExplicitContextTag2::from(OctetStringAsn1::from(ticket_enc_data)),
+            }),
+        })
+    }
+
     /// Performs AS exchange according to the RFC.
     ///
     /// https://www.rfc-editor.org/rfc/rfc4120#section-3.1
     pub fn as_exchange(&self, as_req: AsReq) -> Result<AsRep, KrbError> {
+        (self.validators.as_req)(&as_req);
+
         let KdcReq {
             pvno: _,
             msg_type: _,
@@ -222,7 +309,7 @@ impl KdcMock {
             .get(&username)
             .expect("user's credentials is not found in KDC database");
 
-        KdcMock::validate_timestamp(
+        Self::validate_timestamp(
             creds,
             sname.clone(),
             realm.clone(),
@@ -248,34 +335,6 @@ impl KdcMock {
         let end_time = auth_time + Duration::days(1);
         let realm = Realm::from(IA5String::from_string(self.realm.clone()).unwrap());
 
-        let ticket_enc_part = EncTicketPart {
-            flags: kdc_options.clone(),
-            key: ExplicitContextTag1::from(EncryptionKey {
-                key_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![18])),
-                key_value: ExplicitContextTag1::from(OctetStringAsn1::from(session_key.to_vec())),
-            }),
-            crealm: ExplicitContextTag2::from(realm.clone()),
-            cname: ExplicitContextTag3::from(username.0.clone()),
-            transited: ExplicitContextTag4::from(TransitedEncoding {
-                // the client is unable to check these fields, so we can put any values we want
-                tr_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![0])),
-                contents: ExplicitContextTag1::from(OctetStringAsn1::from(vec![1])),
-            }),
-            auth_time: ExplicitContextTag5::from(KerberosTime::from(GeneralizedTime::from(auth_time))),
-            starttime: Optional::from(None),
-            endtime: ExplicitContextTag7::from(KerberosTime::from(GeneralizedTime::from(end_time))),
-            renew_till: Optional::from(None),
-            caddr: Optional::from(None),
-            authorization_data: Optional::from(None),
-        };
-        let ticket_enc_data = cipher
-            .encrypt(
-                service_key,
-                TICKET_REP,
-                &picky_asn1_der::to_vec(&ticket_enc_part).unwrap(),
-            )
-            .unwrap();
-
         let as_rep_enc_part = EncAsRepPart::from(EncKdcRepPart {
             key: ExplicitContextTag0::from(EncryptionKey {
                 key_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
@@ -289,7 +348,7 @@ impl KdcMock {
             }])),
             nonce: ExplicitContextTag2::from(IntegerAsn1::from(rng.gen::<u32>().to_be_bytes().to_vec())),
             key_expiration: Optional::from(None),
-            flags: ExplicitContextTag4::from(kdc_options.0),
+            flags: ExplicitContextTag4::from(kdc_options.0.clone()),
             auth_time: ExplicitContextTag5::from(KerberosTime::from(GeneralizedTime::from(auth_time))),
             start_time: Optional::from(None),
             end_time: ExplicitContextTag7::from(KerberosTime::from(GeneralizedTime::from(end_time))),
@@ -324,17 +383,15 @@ impl KdcMock {
                 )),
             }])))),
             crealm: ExplicitContextTag3::from(realm.clone()),
-            cname: ExplicitContextTag4::from(username.0),
-            ticket: ExplicitContextTag5::from(Ticket::from(TicketInner {
-                tkt_vno: ExplicitContextTag0::from(IntegerAsn1::from(vec![KERBEROS_VERSION])),
-                realm: ExplicitContextTag1::from(realm),
-                sname: ExplicitContextTag2::from(sname),
-                enc_part: ExplicitContextTag3::from(EncryptedData {
-                    etype: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
-                    kvno: Optional::from(None),
-                    cipher: ExplicitContextTag2::from(OctetStringAsn1::from(ticket_enc_data)),
-                }),
-            })),
+            cname: ExplicitContextTag4::from(username.0.clone()),
+            ticket: ExplicitContextTag5::from(Self::make_ticket(
+                realm,
+                session_key,
+                service_key,
+                kdc_options.0,
+                sname,
+                username.0,
+            )),
             enc_part: ExplicitContextTag6::from(EncryptedData {
                 etype: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
                 kvno: Optional::from(None),
@@ -428,6 +485,8 @@ impl KdcMock {
     ///
     /// https://www.rfc-editor.org/rfc/rfc4120#section-3.3
     pub fn tgs_exchange(&self, tgs_req: TgsReq) -> Result<TgsRep, KrbError> {
+        (self.validators.tgs_req)(&tgs_req);
+
         let KdcReq {
             pvno: _,
             msg_type: _,
@@ -502,34 +561,6 @@ impl KdcMock {
         let auth_time = OffsetDateTime::now_utc();
         let end_time = auth_time + Duration::days(1);
 
-        let ticket_enc_part = EncTicketPart {
-            flags: kdc_options.clone(),
-            key: ExplicitContextTag1::from(EncryptionKey {
-                key_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
-                key_value: ExplicitContextTag1::from(OctetStringAsn1::from(session_key.to_vec())),
-            }),
-            crealm: ExplicitContextTag2::from(realm.clone()),
-            cname: ExplicitContextTag3::from(cname.clone()),
-            transited: ExplicitContextTag4::from(TransitedEncoding {
-                // the client is unable to check these fields, so we can put any values we want
-                tr_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![0])),
-                contents: ExplicitContextTag1::from(OctetStringAsn1::from(vec![1])),
-            }),
-            auth_time: ExplicitContextTag5::from(KerberosTime::from(GeneralizedTime::from(auth_time))),
-            starttime: Optional::from(None),
-            endtime: ExplicitContextTag7::from(KerberosTime::from(GeneralizedTime::from(end_time))),
-            renew_till: Optional::from(None),
-            caddr: Optional::from(None),
-            authorization_data: Optional::from(None),
-        };
-        let ticket_enc_data = cipher
-            .encrypt(
-                &ticket_enc_key,
-                TICKET_REP,
-                &picky_asn1_der::to_vec(&ticket_enc_part).unwrap(),
-            )
-            .unwrap();
-
         let tgs_rep_enc_part = EncTgsRepPart::from(EncKdcRepPart {
             key: ExplicitContextTag0::from(EncryptionKey {
                 key_type: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
@@ -543,7 +574,7 @@ impl KdcMock {
             }])),
             nonce: ExplicitContextTag2::from(IntegerAsn1::from(rng.gen::<u32>().to_be_bytes().to_vec())),
             key_expiration: Optional::from(None),
-            flags: ExplicitContextTag4::from(kdc_options.0),
+            flags: ExplicitContextTag4::from(kdc_options.0.clone()),
             auth_time: ExplicitContextTag5::from(KerberosTime::from(GeneralizedTime::from(auth_time))),
             start_time: Optional::from(None),
             end_time: ExplicitContextTag7::from(KerberosTime::from(GeneralizedTime::from(end_time))),
@@ -566,22 +597,47 @@ impl KdcMock {
             msg_type: ExplicitContextTag1::from(IntegerAsn1::from(vec![TGS_REP_MSG_TYPE])),
             padata: Optional::from(None),
             crealm: ExplicitContextTag3::from(realm.clone()),
-            cname: ExplicitContextTag4::from(cname),
-            ticket: ExplicitContextTag5::from(Ticket::from(TicketInner {
-                tkt_vno: ExplicitContextTag0::from(IntegerAsn1::from(vec![KERBEROS_VERSION])),
-                realm: ExplicitContextTag1::from(realm),
-                sname: ExplicitContextTag2::from(sname),
-                enc_part: ExplicitContextTag3::from(EncryptedData {
-                    etype: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
-                    kvno: Optional::from(None),
-                    cipher: ExplicitContextTag2::from(OctetStringAsn1::from(ticket_enc_data)),
-                }),
-            })),
+            cname: ExplicitContextTag4::from(cname.clone()),
+            ticket: ExplicitContextTag5::from(Self::make_ticket(
+                realm,
+                session_key,
+                &ticket_enc_key,
+                kdc_options.0,
+                sname,
+                cname,
+            )),
             enc_part: ExplicitContextTag6::from(EncryptedData {
                 etype: ExplicitContextTag0::from(IntegerAsn1::from(vec![AES256_CTS_HMAC_SHA1_96 as u8])),
                 kvno: Optional::from(None),
                 cipher: ExplicitContextTag2::from(OctetStringAsn1::from(tgs_rep_enc_data)),
             }),
         }))
+    }
+
+    /// Generates TgtRep based on the incoming [TgtReq].
+    ///
+    /// This method helps to simulate Kerberos U2U auth.
+    pub fn generate_tgt(&self, tgt_req: TgtReq, session_key: &[u8]) -> TgtRep {
+        let sname = tgt_req.server_name.0;
+        let realm = Realm::from(IA5String::from_string(self.realm.clone()).unwrap());
+
+        let service_key = self
+            .keys
+            .get(&UserName(sname.clone()))
+            .expect("service's key must present in KDC database")
+            .to_vec();
+
+        TgtRep {
+            pvno: ExplicitContextTag0::from(IntegerAsn1::from(vec![KERBEROS_VERSION])),
+            msg_type: ExplicitContextTag1::from(IntegerAsn1::from(vec![TGT_REP_MSG_TYPE])),
+            ticket: ExplicitContextTag2::from(Self::make_ticket(
+                realm,
+                session_key.to_vec(),
+                &service_key,
+                KerberosFlags::from(BitString::with_bytes([0; 4])),
+                sname.clone(),
+                sname,
+            )),
+        }
     }
 }
